deps: bump dependencies
