+++ package.json
@@ -1,7 +1,7 @@
 {
 	"name": "atomic-calendar-revive",
-	"version": "7.2.1",
-	"editor_version": "2.2.1",
+	"version": "7.2.2",
+	"editor_version": "2.2.2",
 	"description": "Calendar Card for Home Assistant",
 	"main": "atomic-calendar-revive.js",
 	"scripts": {
@@ -24,28 +24,28 @@
 	"author": "Steven Marks",
 	"license": "ISC",
 	"devDependencies": {
-		"@babel/cli": "^7.17.10",
-		"@babel/core": "^7.18.2",
-		"@babel/plugin-proposal-class-properties": "^7.17.12",
-		"@babel/plugin-proposal-decorators": "^7.17.12",
-		"@babel/plugin-transform-template-literals": "^7.18.2",
-		"@babel/preset-env": "^7.18.2",
-		"@rollup/plugin-babel": "^6.0.0",
-		"@rollup/plugin-commonjs": "^23.0.0",
-		"@rollup/plugin-eslint": "^9.0.0",
-		"@rollup/plugin-json": "^4.1.0",
-		"@rollup/plugin-node-resolve": "^14.1.0",
-		"@typescript-eslint/eslint-plugin": "^5.29.0",
-		"@typescript-eslint/parser": "^5.29.0",
+		"@babel/cli": "^7.20.7",
+		"@babel/core": "^7.20.12",
+		"@babel/plugin-proposal-class-properties": "^7.18.6",
+		"@babel/plugin-proposal-decorators": "^7.20.7",
+		"@babel/plugin-transform-template-literals": "^7.18.9",
+		"@babel/preset-env": "^7.20.2",
+		"@rollup/plugin-babel": "^6.0.3",
+		"@rollup/plugin-commonjs": "^24.0.0",
+		"@rollup/plugin-eslint": "^9.0.2",
+		"@rollup/plugin-json": "^6.0.0",
+		"@rollup/plugin-node-resolve": "^15.0.1",
+		"@typescript-eslint/eslint-plugin": "^5.48.2",
+		"@typescript-eslint/parser": "^5.48.2",
 		"babel-preset-minify": "^0.5.2",
-		"commitizen": "^4.2.4",
+		"commitizen": "^4.2.6",
 		"cz-conventional-changelog": "^3.3.0",
-		"eslint": "^8.18.0",
-		"eslint-config-prettier": "^8.5.0",
-		"eslint-plugin-prettier": "^4.0.0",
-		"prettier": "^2.7.1",
-		"rollup": "^2.75.7",
-		"rollup-plugin-serve": "^2.0.1",
+		"eslint": "^8.32.0",
+		"eslint-config-prettier": "^8.6.0",
+		"eslint-plugin-prettier": "^4.2.1",
+		"prettier": "^2.8.3",
+		"rollup": "^3.10.0",
+		"rollup-plugin-serve": "^2.0.2",
 		"rollup-plugin-terser": "^7.0.2",
 		"rollup-plugin-typescript2": "^0.34.1"
 	},
@@ -54,7 +54,7 @@
 		"@material/mwc-icon-button": "^0.27.0",
 		"@material/mwc-linear-progress": "^0.27.0",
 		"@webcomponents/webcomponentsjs": "^2.6.0",
-		"custom-card-helpers": "git+https://github.com/totaldebug/custom-card-helpers.git",
+		"custom-card-helpers": "^1.8.0",
 		"dayjs": "^1.11.2",
 		"lit": "^2.2.6",
 		"lit-element": "^3.2.0",
